TFHUB-70 update versions and general improvements no behavior changes (#1)
* Remove unnecessary integration test marker

The defaults for surefire and failsafe are sufficient.  Defaults are
unit tests must have "Test" at start or end of class name and IT tests
must have "IT" at start or end of class name.

* Update docs

* Removing need for external redis by using docker

* Add enforcer plugin to warn on dependency issues

This does not fail the build on dependency convergence issues nor for
duplicate classes but only outputs warnings during the build.

* Update versions to latest, exclude non-converging transients

Transients were only excluded when the versions clashed with those
in other jars (when dependency convergence enforcement rules show
warnings).

* Correct ignore file name

* Add config to allow jenkins to sign correctly (passphrase also needed in settings.xml pgp.passphrase property)

* Use OSS parent pom to simplify config required to push to OSS

* Use oss-parent pom standard config rather than custom vnomics values

This requires that settings.xml has user/pass set such as the following:

  <servers>
    <server>
      <id>sonatype-nexus-snapshots</id>
      <username>OssUsername</username>
      <password>OssPassword</password>
    </server>
    <server>
      <id>sonatype-nexus-staging</id>
      <username>OssUsername</username>
      <password>OssPassword</password>
    </server>

  </servers>

* Remove unnecessary config

* Update to latest spring security, old version has known security issues

Github actually flagged this so we may not be using anything in
spring security but without knowing upgrading is the best course of
action.